requirements.txt
django
djangorestframework
django-filter
python-dotenv
